+++ inst/example-scripts/irrad-acq-timeseries.R
@@ -7,10 +7,22 @@ library(ooacquire)
 
 photon_as_default()
 
+# half minute of measurements every 2 seconds
+acq_irrad_interactive(interface.mode = "series",
+                      folder.name = "./inst-not/series",
+                      tot.time.range = 1,
+                      target.margin = .1,
+                      qty.out = "irrad",
+                      HDR.mult = 1L,
+                      seq.settings = list(start.boundary = "second",
+                                          initial.delay = 0,
+                                          step.delay = 2,
+                                          num.steps = 15),
+                      entrance.optics = "cosine")
+
 # one hour of measurements every minute
 # Before starting make sure spectrometer temperature is stable!!
-
-acq_irrad_interactive(interface.mode = "series-attr",
+acq_irrad_interactive(interface.mode = "series",
                       folder.name = "./inst-not/series",
                       tot.time.range = c(5, 10),
                       target.margin = .1,
@@ -30,4 +42,18 @@ acq_irrad_interactive(interface.mode = "auto",
                       qty.out = "irrad",
                       HDR.mult = c(0.3, 1, 3, 10),
                       entrance.optics = "cosine")
+
+# high speed measurements as fast as possible
+acq_irrad_interactive(interface.mode = "series-attr",
+                      folder.name = "./inst-not/series",
+                      tot.time.range = c(0, Inf),
+                      target.margin = .2,
+                      qty.out = "irrad",
+                      HDR.mult = 1L,
+                      seq.settings = list(start.boundary = "second",
+                                          initial.delay = 0,
+                                          step.delay = 0,
+                                          num.steps = 100),
+                      entrance.optics = "cosine")
+